move coordinator logic into coordinator.py
