+++ requirements.txt
@@ -1,1 +1,2 @@
 homeassistant==2023.2.0
+aiohttp~=3.8.1
